+++ src/types.ts
@@ -1,4 +1,4 @@
-import type { Rational } from "../rational";
+import type { Rational } from "./rational";
 
 /**
  * Represents a ratio.
